Clickable lables on checkboxes and minor bugfixes
Checkboxes are now clickable, and I changed the default to only check
L6, L7, and L8… Feel free to change this back if it's important to
process the other ones, but for me, I only care about higher level
stuff.

Bug fixes are related to JS errors that would happen if I used chat.